refactor for single responsibility
