+++ classes/course_files.php
@@ -50,16 +50,6 @@ class course_files {
     protected $filelist = null;
 
     /**
-     * @var null
-     */
-    static protected $licenses = null;
-
-    /**
-     * @var null
-     */
-    static protected $licenscolors = null;
-
-    /**
      * @var string
      */
     protected $filtercomponent;
@@ -78,24 +68,6 @@ class course_files {
      * @var int
      */
     protected $courseid;
-
-    /**
-     * Mapping of file types to possible mime types.
-     * @var array
-     */
-    static protected $mimetypes = array(
-        'document' => array('application/epub+zip', 'application/msword', 'application/pdf',
-            'application/postscript', 'application/vnd.ms-%', 'application/vnd.oasis.opendocument%',
-            'application/vnd.openxmlformats-officedocument%', 'application/vnd.sun.xml%',
-            'application/x-digidoc', 'application/xhtml+xml', 'application/x-javascript',
-            'application/x-latex', 'application/xml', 'application/x-ms%', 'application/x-tex%',
-            'document%', 'spreadsheet', 'text/%'),
-        'image' => array('image/%'),
-        'audio' => array('audio/%'),
-        'video' => array('video/%'),
-        'archive' => array('application/zip', 'application/x-tar', 'application/g-zip',
-            'application/x-rar-compressed', 'application/x-7z-compressed', 'application/vnd.moodle.backup'),
-    );
 
     /**
      * course_files constructor.
@@ -139,8 +111,8 @@ class course_files {
         }
 
         if ($this->filterfiletype === 'other') {
-            $sqlwhere .= ' AND ' . $this->get_sql_mimetype(array_keys(self::$mimetypes), false);
-        } else if (isset(self::$mimetypes[$this->filterfiletype])) {
+            $sqlwhere .= ' AND ' . $this->get_sql_mimetype(array_keys(mimetypes::get_mime_types()), false);
+        } else if (isset(mimetypes::get_mime_types()[$this->filterfiletype])) {
             $sqlwhere .= ' AND ' . $this->get_sql_mimetype($this->filterfiletype, true);
         }
 
@@ -188,7 +160,7 @@ class course_files {
                 $list = array_merge($list, self::$mimetypes[$type]);
             }
         } else {
-            $list = &self::$mimetypes[$types];
+            $list = &mimetypes::get_mime_types()[$types];
         }
 
         if ($in) {
@@ -244,53 +216,6 @@ class course_files {
         $this->components = $componentsall + $this->components;
 
         return $this->components;
-    }
-
-    /**
-     * Remember licences as array
-     *
-     * @return array|null
-     * @throws \coding_exception
-     */
-    public static function get_available_licenses() {
-        global $CFG;
-
-        if (self::$licenses === null) {
-            self::$licenses = array();
-            $a = explode(',', $CFG->licenses);
-            foreach ($a as $license) {
-                self::$licenses[$license] = \get_string($license, 'license');
-            }
-        }
-
-        return self::$licenses;
-    }
-
-    /**
-     * Wraps license name in span element with background color as per plugin settings or
-     * retuns license name if no color set
-     *
-     * @param string $licenseshort short name of a license
-     * @return string full name of the license with HTML
-     * @throws dml_exception|coding_exception
-     */
-    public static function get_license_name_color($licenseshort) {
-        if (self::$licenscolors === null) {
-            self::get_available_licenses();
-            $colorscfg = get_config('local_listcoursefiles', 'licensecolors');
-            $matches = array();
-            preg_match_all('@\s*(\S+)\s*([a-fA-F0-9]{6})\s*@', $colorscfg, $matches, PREG_SET_ORDER);
-            self::$licenscolors = array();
-            foreach ($matches as $m) {
-                self::$licenscolors[$m[1]] = $m[2];
-            }
-        }
-
-        $name = (isset(self::$licenses[$licenseshort])) ? self::$licenses[$licenseshort] : '';
-        if (isset(self::$licenscolors[$licenseshort])) {
-            $name = \html_writer::tag('span', $name, array('style' => 'background-color: #' . self::$licenscolors[$licenseshort]));
-        }
-        return $name;
     }
 
     /**
@@ -444,120 +369,6 @@ class course_files {
     }
 
     /**
-     * Try to get the url for the component (module or course).
-     *
-     * @param object $file
-     * @return null|\moodle_url
-     * @throws moodle_exception
-     */
-    public function get_component_url($file) {
-        if ($file->contextlevel == CONTEXT_MODULE) {
-            if (!empty($this->coursemodinfo->cms[$file->instanceid])) {
-                return $this->coursemodinfo->cms[$file->instanceid]->url;
-            }
-        } else if ($file->contextlevel == CONTEXT_COURSE) {
-            if ($file->component === 'contentbank') {
-                return new \moodle_url('/contentbank/index.php', ['contextid' => $file->contextid]);
-            } else if ($file->filearea === 'section') {
-                return new \moodle_url('/course/view.php', array(
-                    'id' => $this->courseid,
-                    'sectionid' => $file->itemid
-                ));
-            } else {
-                return new \moodle_url('/course/info.php', array(
-                    'id' => $this->courseid
-                ));
-            }
-        }
-
-        return null;
-    }
-
-    /**
-     * Try to get the download url for a file.
-     *
-     * @param array $file
-     * @return null|moodle_url
-     */
-    public function get_file_download_url($file) {
-        switch ($file->component . '#' . $file->filearea) {
-            case 'mod_folder#intro':
-            case 'mod_folder#content':
-            case 'mod_resource#intro':
-            case 'mod_resource#content':
-                return new \moodle_url('/pluginfile.php/'. $file->contextid . '/' . $file->component . '/' .
-                        $file->filearea . '/0' . $file->filepath . $file->filename);
-
-            case 'mod_assign#intro':
-            case 'mod_label#intro':
-                return new \moodle_url('/pluginfile.php/'. $file->contextid . '/' . $file->component . '/' .
-                        $file->filearea . '/' . $file->filepath . $file->filename);
-
-            case 'assignsubmission_file#submission_files':
-            case 'mod_assign#introattachment':
-            case 'mod_data#content':
-            case 'mod_forum#post':
-            case 'mod_forum#attachment':
-            case 'mod_page#content':
-            case 'mod_page#intro':
-            case 'mod_glossary#entry':
-            case 'mod_wiki#attachments':
-            case 'course#section':
-                return new \moodle_url('/pluginfile.php/'. $file->contextid . '/' . $file->component . '/' .
-                        $file->filearea . '/' . $file->itemid . $file->filepath . $file->filename);
-
-            case 'course#legacy':
-                return new \moodle_url('/file.php/'. $this->courseid . $file->filepath . $file->filename);
-        }
-
-        return null;
-    }
-
-    /**
-     * Check if a file with a specific license has expired.
-     *
-     * This checks if a file has been expired because:
-     *  - it is a document (has a particular mimetype),
-     *  - has been provided under a specific license
-     *  - and the expiry date is exceed (in respect of the coure start date and the file creation time).
-     *
-     * The following settings need to be defined in the config.php:
-     *   array $CFG->fileexpirylicenses which licenses (shortnames) expire
-     *   int $CFG->fileexpirydate when do files expire (unix time)
-     *   array $CFG->filemimetypes['document'] mime types of documents
-     *
-     * These adjustments were made by Technische Universität Berlin in order to conform to § 52a UrhG.
-     *
-     * @param stdClass $file
-     * @return boolean whether file is allowed to be delivered to students
-     */
-    public function check_mimetype_license_expiry_date($file) {
-        global $CFG, $COURSE;
-
-        // Check if enabled/configured.
-        if (!isset($CFG->fileexpirydate, $CFG->fileexpirylicenses, $CFG->filemimetypes['document'])) {
-            return true;
-        }
-
-        if (in_array($file->license, $CFG->fileexpirylicenses)) {
-            $isdoc = false;
-            $fmimetype = $file->mimetype;
-            foreach ($CFG->filemimetypes['document'] as $mime) {
-                if ($mime === $fmimetype || (substr($mime, -1) === '%' && strncmp($mime, $fmimetype, strlen($mime) - 1) === 0)) {
-                    $isdoc = true;
-                    break;
-                }
-            }
-            $coursestart = isset($COURSE->startdate) ? $COURSE->startdate : 0;
-            if ($isdoc && $CFG->fileexpirydate > max($coursestart, $file->timecreated)) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    /**
      * Collate an array of available file types
      *
      * @return array
@@ -565,284 +376,11 @@ class course_files {
      */
     public static function get_file_types() {
         $types = array('all' => \get_string('filetype_all', 'local_listcoursefiles'));
-        foreach (self::$mimetypes as $type => $unused) {
+        foreach (mimetypes::get_mime_types() as $type => $unused) {
             $types[$type] = \get_string('filetype_' . $type, 'local_listcoursefiles');
         }
         $types['other'] = \get_string('filetype_other', 'local_listcoursefiles');
         return $types;
     }
 
-    /**
-     * Try to get the name of the file type in the user's lang
-     *
-     * @param string $mimetype
-     * @return \lang_string|string
-     * @throws \coding_exception
-     */
-    public static function get_file_type_translation($mimetype) {
-        foreach (self::$mimetypes as $name => $types) {
-            foreach ($types as $mime) {
-                if ($mime === $mimetype ||
-                        (substr($mime, -1) === '%' && strncmp($mime, $mimetype, strlen($mime) - 1) === 0)) {
-                    return \get_string('filetype_' . $name, 'local_listcoursefiles');
-                }
-            }
-        }
-
-        return $mimetype;
-    }
-
-    /**
-     * Check if the predefined list of mimetypes should be overridden.
-     */
-    public static function check_config_mimetypes() {
-        global $CFG;
-
-        if (isset($CFG->filemimetypes)) {
-            self::$mimetypes = $CFG->filemimetypes;
-        }
-    }
-
-    /**
-     * Checks if embedded files have been used
-     *
-     * @param object $file
-     * @param integer $courseid
-     * @return bool
-     * @throws \dml_exception
-     */
-    public function is_file_used($file, $courseid) {
-        global $DB;
-        $isused = false;
-
-        switch ($file->contextlevel){
-            case '50' : // Course.
-                if ($file->component === 'contentbank') {
-                    $isused = null;
-                } else if ($file->filearea === 'section') {
-                    $section = $DB->get_record('course_sections', ['id' => $file->itemid]);
-                    $isused = $this->is_embedded_file_used($section, 'summary', $file->filename);
-                } else if ($file->filearea === 'overviewfiles') {
-                    $isused = true;
-                } else if ($file->filearea === 'summary') {
-                    $course = $DB->get_record('course', ['id' => $courseid]);
-                    $isused = $this->is_embedded_file_used($course, 'summary', $file->filename);
-                }
-                break;
-            case '70' : // Course module.
-                $modname = str_replace('mod_', '', $file->component);
-                if ($file->filearea === 'intro') {
-                    $sql = 'SELECT m.* FROM {context} ctx
-                            JOIN {course_modules} cm ON cm.id = ctx.instanceid
-                            JOIN {' . $modname . '} m ON m.id = cm.instance
-                            WHERE ctx.id = ?';
-                    $mod = $DB->get_record_sql($sql, [$file->contextid]);
-                    $isused = $this->is_embedded_file_used($mod, 'intro', $file->filename);
-                } else {
-                    $fn = 'is_file_used_' . $modname;
-                    if (is_callable(array($this, $fn))) {
-                        $isused = call_user_func(array($this, $fn), $file);
-                    } else {
-                        $isused = null;
-                    }
-                }
-                break;
-            default :
-                $isused = null;
-        }
-        return $isused;
-    }
-
-    /**
-     * Additional file used checks for the Assignment activity module
-     * @param object $file
-     * @return bool | null
-     */
-    private function is_file_used_assign($file) {
-        // File areas = intro, introattachment.
-        if ($file->filearea === 'introattachment') {
-            return true;
-        }
-    }
-
-    /**
-     * Additional file used checks for the Book resource module
-     * @param object $file
-     * @return bool | null
-     * @throws \dml_exception
-     */
-    private function is_file_used_book($file) {
-        // File areas = intro, chapter.
-        global $DB;
-        if ($file->filearea === 'chapter') {
-            $chapter = $DB->get_record('book_chapters', ['id' => $file->itemid]);
-            $isused = $this->is_embedded_file_used($chapter, 'content', $file->filename);
-            return $isused;
-        }
-    }
-
-    /**
-     * Additional file used checks for the Database activity module
-     * @param object $file
-     * @return bool | null
-     * @throws \dml_exception
-     */
-    private function is_file_used_data($file) {
-        // File areas = intro, content.
-        global $DB;
-        if ($file->filearea === 'content') {
-            $sql = 'SELECT * FROM {data_content} dc
-                    JOIN {data_fields} df ON df.id = dc.fieldid
-                    WHERE dc.id = ?';
-            $data = $DB->get_record_sql($sql, [$file->itemid]);
-            if ($data->type !== 'textarea' ||
-                false !== strpos($data->content, '@@PLUGINFILE@@/' . rawurlencode($file->filename))) {
-                return true;
-            } else {
-                return false;
-            }
-        }
-    }
-
-    /**
-     * Additional file used checks for the Feedback activity module
-     * @param object $file
-     * @return bool | null
-     * @throws \dml_exception
-     */
-    private function is_file_used_feedback($file) {
-        // File areas = intro, item, page_after_submit.
-        global $DB;
-        if ($file->filearea === 'item') {
-            $item = $DB->get_record('feedback_item', ['id' => $file->itemid]);
-            $isused = $this->is_embedded_file_used($item, 'presentation', $file->filename);
-            return $isused;
-        } else if ($file->filearea = 'page_after_submit') {
-            $sql = 'SELECT m.* FROM {feedback} m
-                    JOIN {course_modules} cm ON cm.instance = m.id
-                    JOIN {context} ctx ON ctx.instanceid = cm.id
-                    WHERE ctx.id = ?';
-            $feedback = $DB->get_record_sql($sql, [$file->contextid]);
-            $isused = $this->is_embedded_file_used($feedback, 'page_after_submit', $file->filename);
-            return $isused;
-        }
-    }
-
-    /**
-     * Additional file used checks for Folder resource module
-     * @return bool
-     */
-    private function is_file_used_folder() {
-        // File areas = intro, content.
-        return true;
-    }
-
-    /**
-     * Additional file used checks for the Forum activity module
-     * @param object $file
-     * @return bool | null
-     * @throws \dml_exception
-     */
-    private function is_file_used_forum($file) {
-        // File areas = intro, post.
-        global $DB;
-        if ($file->filearea === 'post') {
-            $post = $DB->get_record('forum_posts', ['id' => $file->itemid]);
-            $isused = $this->is_embedded_file_used($post, 'forum_posts', $file->filename);
-            return $isused;
-        } else if ($file->filearea === 'attachment') {
-            return true;
-        }
-    }
-
-    /**
-     * Additional file used checks for the H5P activity module
-     * @param object $file
-     * @return bool
-     */
-    private function is_file_used_h5pactivity() {
-        // File areas = intro, package.
-        return true;
-    }
-
-    /**
-     * Additional file used checks for the Page resource module
-     * @param object file
-     * @return bool | null
-     * @throws \dml_exception
-     */
-    private function is_file_used_page($file) {
-        // File areas = intro, content.
-        global $DB;
-        if ($file->filearea === 'content') {
-            $sql = 'SELECT m.* FROM {page} m
-                    JOIN {course_modules} cm ON cm.instance = m.id
-                    JOIN {context} ctx ON ctx.instanceid = cm.id
-                    WHERE ctx.id = ?';
-            $page = $DB->get_record_sql($sql, [$file->contextid]);
-            $isused = $this->is_embedded_file_used($page, 'content', $file->filename);
-            return $isused;
-        }
-    }
-
-    /**
-     * Additional file used checks for the Resource module
-     * @return bool
-     */
-    private function is_file_used_resource() {
-        // File areas = intro, content.
-        return true;
-    }
-
-    /**
-     * Test if a file is embbeded in text
-     *
-     * @param object $record
-     * @param string $field
-     * @param string $filename
-     * @return bool|null
-     */
-    private function is_embedded_file_used($record, $field, $filename) {
-        if ($record) {
-            return is_int(strpos($record->$field, '@@PLUGINFILE@@/' . rawurlencode($filename)));
-        } else {
-            return null;
-        }
-    }
-
-
-    /**
-     * Creates the URL for the editor where the file is added
-     *
-     * @param object $file
-     * @return \moodle_url|string
-     * @throws \dml_exception
-     * @throws \moodle_exception
-     */
-    public function get_edit_url($file) {
-        global $DB;
-        $url = '';
-        switch ($file->contextlevel){
-            case '50' :
-                if ($file->filearea === 'section') {
-                    $url = new \moodle_url('/course/editsection.php?', ['id' => $file->itemid]);
-                } else if ($file->filearea === 'overviewfiles' || $file->filearea === 'summary') {
-                    $url = new \moodle_url('/course/edit.php?', ['id' => $this->courseid]);
-                }
-                break;
-            case '70' :
-                if ($file->filearea !== 'intro') { // Just checking description for now.
-                    break;
-                }
-                $sql = 'SELECT cm.* FROM {context} ctx
-                        JOIN {course_modules} cm ON cm.id = ctx.instanceid
-                        WHERE ctx.id = ?';
-                $mod = $DB->get_record_sql($sql, [$file->contextid]);
-                $url = new \moodle_url('/course/modedit.php?', ['update' => $mod->id]);
-                break;
-            default :
-                $url = '';
-        }
-        return $url;
-    }
 }
